webapi/webapi-devicecapabilities-w3c-tests/devicecapabilities/SystemDisplayEvent_DisplayUnit-manual.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Xu, Kang <kangx.xu@intel.com>

-->

<meta charset='utf-8'>
<title>DeviceCapabilities Test: SystemDisplayEvent_DisplayUnit-manual</title>
<link rel="author" title="Intel" href="http://www.intel.com/">
<link rel="help" href="http://www.w3.org/2012/sysapps/device-capabilities/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<meta name="flags" content="interact">
<meta name="assert" content="Check if the readonly attribute SystemDisplayEvent.DisplayUnit exists and typeof object">
<p>Connect a new display to the system</p>
<p>Test passes if "PASS" result appears after you connect a new display to the system</p>
<div id="log"></div>
<script>
var t = async_test("Check if the readonly attribute SystemDisplayEvent.DisplayUnit exists and typeof object", {timeout: 40000});

t.step(function () {
  assert_true(!!system, "The Navigator support system");

  system.addEventListener("displayconnect", function (e) {
    t.step(function () {
      assert_true("display" in e, "The SystemDisplayEvent.DisplayUnit exists");
      assert_equals(typeof e.display, "object", "The type of SystemDisplayEvent.DisplayUnit");
      var value = e.display;
      e.display= null;
      assert_equals(e.display, value, "The SystemDisplayEvent.DisplayUnit");
    });
    t.done();
  });
});
</script>
